+++ pymc3/tests/test_plots.py
@@ -6,7 +6,7 @@ import pymc3 as pm
 from .checks import close_to
 
 from .models import multidimensional_model, simple_categorical
-from ..plots import traceplot, forestplot, autocorrplot, plot_posterior, energyplot, densityplot
+from ..plots import traceplot, forestplot, autocorrplot, plot_posterior, energyplot, densityplot, scatterplot
 from ..plots.utils import make_2d
 from ..step_methods import Slice, Metropolis
 from ..sampling import sample
@@ -66,6 +66,7 @@ def test_plots_multidimensional():
     forestplot(trace)
     densityplot(trace)
 
+
 @pytest.mark.xfail(condition=(theano.config.floatX == "float32"), reason="Fails on GPU due to cores=2")
 def test_multichain_plots():
     model = build_disaster_model()
@@ -118,3 +119,15 @@ def test_plots_transformed():
     assert autocorrplot(trace, plot_transformed=True).shape == (2, 2)
     assert plot_posterior(trace).numCols == 1
     assert plot_posterior(trace, plot_transformed=True).shape == (2, )
+
+def test_scatterplot():
+    with pm.Model() as model:
+        a = pm.Normal('a', shape=2)
+        c = pm.HalfNormal('c', shape=2)
+        b = pm.Normal('b', a, c, shape=2)
+        d = pm.Normal('d', 100, 1)
+        trace = pm.sample(1000)
+
+    scatterplot(trace)
+    scatterplot(trace, hexbin=True, plot_transformed=True)
+    scatterplot(trace, sub_varnames=['a_0', 'c_0', 'b_1'])